Added a Model Checker for Computation Tree Logic (CTL), see http://en.wikipedia.org/wiki/Computation_tree_logic
Also added a factory for instantiating ModelCheckers. ModelCheckers have a method
   void ModelCheck(Graph &G, std::istream formula)
which uses the purely virtual method
   Formula* Parse(std::istream)
(using flex+bison to write this is the intention)
to get a Formula* instance, which has a purely virtual method
   void Interpret(Graph& G).

What the formula does on G is up to the formula. Something like an SQL-equivalent for editing graphs would be nice.

git-svn-id: svn://svn.code.sf.net/p/opengraphtheory/code/Trunk@93 7d856203-d9e4-4e18-aee0-bf315c1b057e