fixed the issue mentioned in the previous commit.
Further, OGoCreator now goes to a nice initial zoom (so that the physical location of things is equal to their
logical coordinates - if a vertex is at location (x,y), then it is drawn at x cm left, y cm down. this is perfect,
because in gxl, 1 unit stands for 1 cm).
it should be mentioned that OGoCreator queries the screens resolution to determine the zoom precisely.
No tinkering with "Magic-Numbers". The grid is now also displayed in 1cm increments.

I find this very nice :-)

git-svn-id: svn://svn.code.sf.net/p/opengraphtheory/code/Trunk@145 7d856203-d9e4-4e18-aee0-bf315c1b057e