The ogtdisplay program now exits gracefully when the X button is clicked (before, DisplayWindow tried to get the same semaphore twice when the X button was clicked)
also, some glitches in the EdgeIterator constructors were fixed - now EdgeIterators can (as they should) be used as values for maps etc. You can now do something like
  map<Graph::VertexIterator, Graph::EdgeIterator> EdgeToPredecessor;
  EdgeToPredecessor[v] = e;
before, this would have caused a SEGFAULT, because map::operator[] inserts an empty value (constructed by the empty constructor -> this caused the problem) for the
key v and returns it as LVALUE (so it can e.g. be overwritten with =e; )

git-svn-id: svn://svn.code.sf.net/p/opengraphtheory/code/Trunk@154 7d856203-d9e4-4e18-aee0-bf315c1b057e